Set Default Location
I was placing this file on a network share and assumed that the file is was going to open and run in the directory it was in. That was wrong. When the file was on the the desktop, it ran from the desktop. When it was on the a network share, it ran from C$\WIndows\System. So, I added a $HomeDirectory Variable. This needs to be filled out before running. 

I need to implement a default location if this variable is not declared. 

I also REALLY want to just leave the file location I need in this but I do not feel comfortable doing that as that would be a network share. And nobody needs to know the details of my environment.